Код разбит на модули const.js, data.js, util.js, main.js
